feat: 모먼트 상세 페이지 UI 개선
- 피그마와 같이 개선
- 참여한 사람일 때만 모먼트 정보 수정 UI 띄우기
