+++ src/entities/moment/ui/MomentPlaceHeader/index.module.scss
@@ -7,7 +7,7 @@
 }
 
 .title {
-  @include font-text(heading-medium);
+  @include font-text(heading-large);
   color: var(--text-primary);
   @include util-text-overflow-ellipsis;
 }
