--- src/features/moment/ui/MomentInformationHeader/index.module.scss
@@ -1,13 +1,0 @@
-@use '@/shared/styles/global.scss' as *;
-
-.container {
-  @include flex-row(between, center);
-  margin-bottom: $space-md;
-  gap: $space-md;
-}
-
-.title {
-  @include font-text(heading-medium);
-  color: var(--text-primary);
-  @include util-text-overflow-ellipsis;
-}
