+++ src/pages/MomentDetailPage/index.tsx
@@ -1,6 +1,5 @@
 import MomentDetailHeader from '@/features/moment/ui/MomentDetailHeader';
 import MomentInformation from '@/entities/moment/ui/MomentInformation';
-import MomentInformationHeader from '@/features/moment/ui/MomentInformationHeader';
 import MomentPlaceHeader from '@/entities/moment/ui/MomentPlaceHeader';
 import MomentPlaceRow from '@/features/moment/ui/MomentPlaceRow';
 import styles from './index.module.scss';
@@ -8,15 +7,14 @@ import styles from './index.module.scss';
 function MomentDetailPage() {
   return (
     <main className={styles.container}>
-      <MomentDetailHeader />
-      <div>
-        <MomentInformationHeader />
+      <section>
+        <MomentDetailHeader />
         <MomentInformation />
-      </div>
-      <div>
+      </section>
+      <section>
         <MomentPlaceHeader />
         <MomentPlaceRow />
-      </div>
+      </section>
     </main>
   );
 }
